spiMongodb/spiDjango/requirements.txt
Django
django-mongoengine
six
pymongo[srv]
djangorestframework
django-rest-framework-mongoengine
Pillow
django-cors-headers
psycopg2
blinker
django-rest-passwordreset
